+++ src/pages/api-documentation/api.module.scss
@@ -1,8 +1,27 @@
 @use '../../variables.module.scss' as *;
 
-.hoverLink {
-  color: $primary;
-  font-weight: $semi-bold-weight;
+div.aboutPageWrapper {
+  margin-top: 2rem;
+  padding-top: 0;
+}
+
+.customTop {
+  padding-top: 26.5px;
+}
+
+.section section {
+  &:first-child {
+    padding-top: 0;
+  }
+
+  &.noBullets,
+  .noBullets {
+    list-style: none;
+  }
+  &:last-child,
+  &.lastChild {
+    padding-bottom: 7.25rem;
+  }
 }
 
 .content {
@@ -10,122 +29,56 @@
   margin-bottom: 3.75rem;
 }
 
-div.apiPageWrapper {
-  margin-top: 2rem;
-  padding-top: 0;
-}
-
-code.code {
-  padding: 10px;
-  display: inline-block;
-  width: 100%;
-  word-break: break-all;
-  &.marginBottom {
-    margin-bottom: 0.5rem;
+.list {
+  li {
+    margin-bottom: 1.5rem;
+  }
+  &.noBullets {
+    list-style: none;
   }
 }
 
-.beautified {
-  width: 100%;
-  pre {
-    white-space: pre-wrap;
+.title {
+  &,
+  h2 {
+    margin-top: 0;
+    margin-bottom: 0;
+    @include headingStyle2-1;
   }
+  h3 {
+    margin-top: 0;
+    @include headingStyle3;
+  }
+  margin-top: 0;
 }
 
-#toc {
-  top: 4rem;
-
-  &.tocClosed {
-    display: none;
-  }
-  &.tocOpen {
-    display: block;
-  }
-}
-.apiPageSpacer {
-  margin-bottom: 30rem;
+.activeLink {
+  color: $primary;
+  font-weight: $semi-bold-weight;
+  border-left: 6px solid $primary !important;
+  padding-left: calc(1rem - 6px);
 }
 
-//#content {
-//  &.tocClosed {
-//    display: block;
-//  }
-//  &.tocOpen {
-//    display: none;
-//  }
-//}
-
-.link {
+.linkClass {
   display: block;
-  box-shadow: 0 1px 0 0 $border-color;
-  padding: 0.6875rem 1rem;
-  line-height: 1rem;
-  cursor: pointer;
-
-  &.headingLevel3 {
-    padding-left: 2rem;
-  }
-
-  &.headingLevel4 {
-    padding-left: 3rem;
-  }
-
-  &.activeLink {
-    color: $primary;
-    font-weight: $semi-bold-weight;
-    border-left: 6px solid $primary;
-    padding-left: calc(1rem - 6px);
-
-    &.headingLevel2 {
-      padding-left: calc(1rem - 6px);
-    }
-
-    &.headingLevel3 {
-      padding-left: calc(2rem - 6px);
-    }
-
-    &.headingLevel4 {
-      padding-left: calc(3rem - 6px);
-    }
-  }
-
-  &:hover {
-    color: $primary;
-    font-weight: $semi-bold-weight;
-  }
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
 }
 
-.sectionBreak section:last-of-type {
-  padding-bottom: 100px;
-}
-
-.sectionBreak section:first-child {
-  padding-top: 0;
+.hoverLink {
+  color: $primary;
+  font-weight: $semi-bold-weight;
 }
 
 @media screen and (min-width: $breakpoint-lg) {
-  .apiPageWrapper {
+  .aboutPageWrapper {
     display: flex;
   }
 
   #content {
     max-width: 847px;
     width: calc(100% - 260px - 2rem);
-
-    &.tocOpen {
-      display: block;
-    }
+    padding-bottom: 15rem;
   }
-
-  #toc {
-    &.tocClosed {
-      display: block;
-    }
-  }
-}
-
-.sectionList {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
 }
